# openapi.yaml
openapi: 3.1.0
info:
  title: MinusPod API
  description: |
    REST API for managing podcast feeds with automatic ad detection and removal.


    This API allows you to:
    - Manage podcast feed subscriptions
    - View and manage processed episodes
    - Configure ad detection settings including AI model selection
    - Monitor system status and trigger cleanup operations
    - Manage cross-episode ad patterns with network and podcast scope
    - Submit corrections to improve ad detection accuracy
  version: 2.0.0
  contact:
    name: MinusPod
  license:
    name: MIT
servers:
  # Same-origin relative URL. Swagger UI served from /api/v1/docs uses
  # whatever host the browser is on, so "Try it out" always hits the
  # correct server: localhost in dev, the production domain in prod,
  # a Tailscale/Cloudflare URL on a tunnel, etc. No hardcoded host.
  - url: /api/v1
    description: Same-origin (current server)
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: '{baseUrl}/api/v1'
    description: Custom server
    variables:
      baseUrl:
        default: http://localhost:8000
        description: Base URL of the podcast server
...
paths:
...
  /history:
    get:
      tags:
        - History
      summary: Get processing history
      description: Returns processing history with pagination and filtering
      operationId: getProcessingHistory
      parameters:
        - name: status
          in: query
          description: Filter by processing status
          schema:
            type: string
            enum: [all, completed, failed]
            default: all
        - name: podcast_slug
          in: query
          description: Filter by podcast slug
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of entries to return
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          description: Number of entries to skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: page
          in: query
          description: Page number (1-based). Alternative to offset. If both provided, page takes precedence.
          schema:
            type: integer
            minimum: 1
        - name: sort_by
          in: query
          description: Column to sort by
          schema:
            type: string
            enum: [processed_at, podcast_title, episode_title, processing_duration_seconds, ads_detected, reprocess_number, status, llm_cost]
            default: processed_at
        - name: sort_dir
          in: query
          description: Sort direction
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Processing history
          content:
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProcessingHistoryEntry'
                  total:
                    type: integer
                    description: Total number of history entries matching the filter
                  totalPages:
                    type: integer
                    description: Total number of pages available
                  limit:
                    type: integer
                  offset:
                    type: integer
                  page:
                    type: integer
                    description: Current page number (1-based)
...
  /history/export:
    get:
      tags:
        - History
      summary: Export processing history
      description: Export processing history as CSV or JSON file
      operationId: exportProcessingHistory
      parameters:
        - name: format
          in: query
          description: Export format
          schema:
            type: string
            enum: [csv, json]
            default: csv
        - name: status
          in: query
          description: Filter by processing status
          schema:
            type: string
            enum: [all, completed, failed]
            default: all
        - name: podcast_slug
          in: query
          description: Filter by podcast slug
          schema:
            type: string
      responses:
        '200':
          description: Exported history file
          content:
            text/csv:
              schema:
                type: string
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProcessingHistoryEntry'
...
  /settings/providers/{provider}:
    put:
      tags:
        - Settings
      summary: Update provider configuration
      operationId: updateProvider
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
            enum: [anthropic, openai, openrouter, whisper, ollama]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                apiKey:  { type: string, nullable: true, description: 'Omit to leave unchanged; null or empty to clear' }
                baseUrl: { type: string }
                model:   { type: string }
      responses:
        '200': { description: Updated status, content: { application/json: { schema: { $ref: '#/components/schemas/ProviderStatus' } } } }
        '400': { description: Invalid input (e.g., baseUrl failed SSRF validation) }
        '404': { description: Unknown provider }
        '409': { description: provider_crypto_unavailable (MINUSPOD_MASTER_PASSPHRASE unset) }
    delete:
      tags:
        - Settings
      summary: Clear stored provider key
      description: Removes the DB-stored key so env-var fallback (if any) takes over.
      operationId: clearProvider
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
            enum: [anthropic, openai, openrouter, whisper, ollama]
      responses:
        '200': { description: Cleared }
        '404': { description: Unknown provider }
  /settings/providers/rotate-passphrase:
    post:
      tags:
        - Settings
      summary: Rotate the master passphrase used to encrypt provider keys
      description: |
        Re-encrypts every stored provider key under a new passphrase + new
        salt in a single SQLite transaction. The caller MUST update
        `MINUSPOD_MASTER_PASSPHRASE` in the container environment to the new
        value before the next restart, otherwise stored keys will not decrypt.
      operationId: rotateMasterPassphrase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [oldPassphrase, newPassphrase]
              properties:
                oldPassphrase: { type: string, description: 'Current MINUSPOD_MASTER_PASSPHRASE value' }
                newPassphrase: { type: string }
      responses:
        '200':
          description: Rotation complete
          content:
            application/json:
              schema:
                type: object
                properties:
                  rotated: { type: integer, description: 'Number of secret rows re-encrypted' }
        '400': { description: 'Missing or invalid input (e.g., current passphrase mismatch, new equals old)' }
        '409': { description: 'provider_crypto_unavailable (MINUSPOD_MASTER_PASSPHRASE unset)' }
        '500': { description: 'Rotation failed (rolled back)' }
  /settings/providers/{provider}/test:
    post:
      tags:
        - Settings
      summary: Test a provider connection
      operationId: testProvider
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
            enum: [anthropic, openai, openrouter, whisper, ollama]
      responses:
        '200':
          description: Test result (always 200; see `ok` field)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:    { type: boolean }
                  error: { type: string }
...
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: |
        Flask session cookie issued by ``POST /auth/login``. Every mutating
        endpoint additionally requires a matching ``X-CSRF-Token`` header
        (double-submit). Public endpoints (``/auth/login``, ``/auth/logout``,
        ``/auth/status``, ``/health``, ``/health/live``) declare ``security: []``.
...
  schemas:
...
    ProviderStatus:
      type: object
      properties:
        configured: { type: boolean }
        source:
          type: string
          enum: [db, env, none]
        baseUrl: { type: string }
        model:   { type: string }
...
    ProcessingHistoryEntry:
      type: object
      properties:
        id:
          type: integer
          description: History entry ID
        podcast_id:
          type: integer
          description: Podcast ID
        podcast_slug:
          type: string
          description: Podcast slug
        podcast_title:
          type: string
          description: Podcast title
        episode_id:
          type: string
          description: Episode ID
        episode_title:
          type: string
          description: Episode title
        processed_at:
          type: string
          format: date-time
          description: When processing completed
        processing_duration_seconds:
          type:
            - number
            - 'null'
          description: Processing duration in seconds
        status:
          type: string
          enum: [completed, failed]
          description: Processing result status
        ads_detected:
          type: integer
          description: Number of ads detected
        error_message:
          type:
            - string
            - 'null'
          description: Error message if processing failed
        reprocess_number:
          type:
            - integer
            - 'null'
          description: Reprocess number (null for first processing)
        inputTokens:
          type: integer
          description: LLM input tokens used for this processing run
        outputTokens:
          type: integer
          description: LLM output tokens used for this processing run
        llmCost:
          type: number
          description: Estimated LLM cost in USD for this processing run
...